@@ -1,10 +1,9 @@
 package com.quchen.flashcard;
 
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AppCompatActivity;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.widget.FrameLayout;
 
 import java.util.ArrayList;
 import java.util.Collections;
